- Rename mp_entry to ap_entry - Create and use cpu_constants.h instead of hardcoded values - Make early boot code work on old processors - Detect old processor and print message - Fix some obsolete uint16_t use - Let libgcc do 64 bit divides - Use double-underscored attribute names - Add configure stuff for stack check and rtti - Add patch helper for self modifying code - Massively improve APIC initialization and IRQ dispatch - Protect CMOS with a lock - Clean up cpu_wait_* functions - Rename MSR_* constants to CPU_MSR_* - Clean up CPU initialization - Align TSS - Store FSBASE and GSBASE in thread - Stop saving FSBASE and GSBASE in context switch - Clean up handling of segment registers in context switch - Reorder format_flag_info_t to avoid padding in the middle of the structure - Initialize IDT and TSS earlier, in startup assembly code - Rearrange interrupt order - simplifies dispatch - Add segment register perf test in isr (temporary) - Explicitly call PIC dispatcher for PIC IRQs - Fix bootloader and use smaller alignments in kernel linker script - Reset AHCI controller in initialization - Improvements in NVME driver (still not passing stress though) - Add helper to configure MSI or pin IRQ, to reduce duplication in every driver - Split iterator and bus/slot/function structure, to avoid keeping unnecessary iterator state in driver data - Add more error handling in filesystem mounts - Add numeric_limits - Add override final to FS_BASE_IMPL macro - Use self modifying code to unconditionally call the appropriate FPU/ SSE/AVX context save - Disable PNG code for now - Fix some structure padding

+++ boot/farptr.h
@@ -36,7 +36,7 @@ static __always_inline void far_zero(far_ptr_t dest, uint16_t paragraphs)
 }
 
 static __always_inline void far_copy(far_ptr_t dest,
-              far_ptr_t src, uint16_t size)
+              far_ptr_t src, size_t size)
 {
     void *dest_addr = far_to_ptr(dest);
     void *src_addr = far_to_ptr(src);
