+++ boot/progressbar.h
@@ -1,5 +1,4 @@
 #pragma once
 #include "types.h"
 
-void progress_bar_draw(uint16_t top, uint16_t left,
-                       uint16_t right, uint16_t percent);
+void progress_bar_draw(int top, int left, int right, int percent);
